requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
flask==3.0.3
anthropic==0.34.0
pdfplumber==0.11.4
pypdf==4.3.1
fastapi==0.115.0
uvicorn[standard]==0.30.0
pyjwt==2.9.0
python-multipart==0.0.9
